Extract auth0 api client
