+++ auth0-lock.js
@@ -2,6 +2,7 @@ var { NativeModules, Platform } = require('react-native');
 var LockModule = NativeModules.Auth0LockModule;
 
 const VERSION = require('./version');
+const Auth0 = require('./auth0');
 
 class Auth0Lock {
   constructor(options) {
@@ -14,6 +15,8 @@ class Auth0Lock {
         libraryVersion: VERSION
       };
       this.nativeIntegrations = options.integrations;
+      const auth0 = new Auth0(domain);
+      this.authentication = auth0.authentication(clientId);
     } else {
       throw "Must supply clientId & domain";
     }
@@ -48,63 +51,11 @@ class Auth0Lock {
   }
 
   delegation(options) {
-    let { clientId, domain } = this.lockOptions;
-    if (!domain.startsWith("http")) {
-      domain = `https://${domain}`;
-    }
-
-    let payload = {
-      "client_id": clientId,
-      "grant_type": "urn:ietf:params:oauth:grant-type:jwt-bearer",
-    };
-
-    let token = options.refreshToken || options.idToken;
-    if (token == null) {
-        return Promise.reject("must supply either a refreshToken or idToken");
-    }
-
-    let attrName = "refresh_token";
-    if (options.refreshToken == null) {
-      attrName = "id_token";
-    }
-
-    payload[attrName] = token;
-
-    if (options.apiType != null) {
-      payload["api_type"] = options.apiType;
-    }
-
-    if (options.target != null) {
-      payload["target"] = options.target;
-    }
-
-    if (options.scope != null) {
-      payload["scope"] = options.scope;
-    }
-
-    return fetch(`${domain}/delegation`, {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(payload)
-    })
-    .then(response => response.json());
+    return this.authentication.delegation(options);
   }
 
   refreshToken(refreshToken, options) {
-    const delegationOptions = Object.assign({}, options);
-    delegationOptions.refreshToken = refreshToken;
-    delegationOptions.apiType = "app";
-    return this.delegation(delegationOptions)
-    .then(json => {
-      return {
-        idToken: json.id_token,
-        expiresIn: json.expires_in,
-        tokenType: json.token_type
-      };
-    });
+    return this.authentication.refreshToken(refreshToken, options);
   }
 }
