+++ auth0.js
@@ -1,0 +1,20 @@
+var Authentication = require('./authentication-api');
+
+class Auth0 {
+  constructor(domain) {
+    if (domain == null) {
+      throw "Must supply a valid Auth0 domain";
+    }
+    let baseUrl = domain
+    if (!domain.startsWith("http")) {
+      baseUrl = `https://${domain}`;
+    }
+    this.baseUrl = baseUrl
+  }
+
+  authentication(clientId) {
+    return new Authentication(clientId, this.baseUrl);
+  }
+}
+
+module.exports = Auth0;
